backend-test/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.1.1
langchain==0.3.26
langchain-community==0.3.27
langchain-huggingface==0.3.0
langchain-anthropic==0.3.4
anthropic>=0.41.0
chromadb==1.0.15
sentence-transformers==3.3.1